Fix warnings from Test::Pod
Test::Pod flags a warning "Expected text after =item, not a number."
for the POD list with items 1, 2, 3 etc in Datebook.pm. This is
because the list starts with 0 so it assumes it is a definitiion list.

Looking at the discussion in
p5p (http://www.nntp.perl.org/group/perl.perl5.porters/2013/02/msg199359.html)
the suggested workaround is to add Z<> after the numeric items.